Calculate subtotal of each item
Subtotal is simply price of item per quantity.
It can be collected looping through each item of the order and sumed up at the end.
The list of product codes needs now to contain uniq product codes, to avoid miscalculations.
Spec still fails though because the offer
on green tea is not yet applied.