Starting logic for applying discounts
In the `total` method, when looping through
the order list, we can check if the quantity of an order item is greater than or equal the minimum quantity needed: if yes, the applicable discount (`discounted_quantity`) is calculated. 
This makes the first case  pass. 
However, this logic will have to be modified for discounts on price and not on quantity.