Discount for strawberries
When the conditions for applying a discount to a specific item exist, we can still think of the subtotal for each item as a `price_per(item) * quantity_per(item)`.
This means dynamically assigning the correct price and quantity, based on the existence of a `discounted_price` or `discounted_quantity` for the specific order item.
If there is no `discounted_price`, the price will fallback to the standard price. If there is no discounted quantity, this will be simply counted as 0.
Now also the second scenario passes.